requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0
django-allauth==0.58.2
django-cors-headers==4.3.1
django-filter==23.5
django-rest-auth==0.9.5
djangorestframework==3.14.0
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
